requirements.txt
tqdm
pycairo
datasets
scikit-learn
transformers
datasets
jiwer
torch
pillow
gradio==3.50
pandas